Fixed some stuff on question application tier

+++ Elearn-ApplicationTier/Elearn.Application/LogicContracts/IQuestionLogic.cs
@@ -7,6 +7,6 @@ public interface IQuestionLogic
     Task<Question> CreateQuestionAsync(Question question);
     Task<Question> GetQuestionByUrlAsync(string url);
     Task<List<Question>> GetQuestionsAsync();
-    Task<List<Question>> GetQuestionByUserIdAsync(int userId);
+    Task<List<Question>> GetQuestionByUserIdAsync(long userId);
     Task<List<Question>> GetQuestionsAsync(int pageNumber, int pageSize);
 }
